+++ src/Hasklepias/Reexports.hs
@@ -31,12 +31,15 @@ module Hasklepias.Reexports (
     , module Data.Function
     , module Data.List
     , module Data.List.NonEmpty
+    , module M
     , module Data.Ord
     , module Data.Proxy
     , module Set
     , module Data.Time.Calendar 
     , module Data.Text
+    , module Data.Traversable 
     , module Data.Tuple
+    , module Data.Tuple.Curry
 
     , module IntervalAlgebra
     , module IntervalAlgebra.IntervalUtilities
@@ -45,12 +48,17 @@ module Hasklepias.Reexports (
     , module Safe
     , module Flow
     , module Witherable
+    , setFromList 
+    , mapFromList
+    , mapToList
 ) where
 
 import safe GHC.Num                         ( Integer(..)
                                             , Num(..)
-                                            , Natural(..) )
-import safe GHC.Real                        ( Integral(..), toInteger )
+                                            , Natural(..)
+                                            , naturalToInt )
+import safe GHC.Real                        ( Integral(..)
+                                            , toInteger )
 import safe GHC.Generics                    ( Generic )
 import safe GHC.Show                        ( Show(..) )
 import safe GHC.TypeLits                    ( KnownSymbol(..)
@@ -70,7 +78,8 @@ import safe Data.Either                     ( Either(..))
 import safe Data.Eq                         ( Eq, (==))
 import safe Data.Foldable                   ( Foldable(..)
                                             , minimum
-                                            , maximum )
+                                            , maximum
+                                             )
 import safe Data.Functor.Contravariant      ( Contravariant(contramap)
                                             , Predicate(..) )
 import safe Data.Function                   ( (.), ($), const, id, flip )
@@ -80,12 +89,20 @@ import safe Data.List                       ( all
                                             , map
                                             , length
                                             , null
+                                            , zip
                                             , zipWith
+                                            , unzip
                                             , replicate
                                             , transpose
                                             , sort
-                                            , (++) )
+                                            , (++)
+                                            , scanl1
+                                            , scanl' )
 import safe Data.List.NonEmpty              ( NonEmpty(..) )
+import safe qualified Data.Map.Strict as M  ( Map(..)
+                                            , toList
+                                            , fromList
+                                            , fromListWith )
 import safe Data.Maybe                      ( Maybe(..),
                                               maybe,
                                               isJust,
@@ -101,7 +118,8 @@ import safe Data.Ord                        ( Ord(..)
                                             , Ordering(..)
                                             , max, min )
 import safe Data.Proxy                      ( Proxy(..) )
-import safe Data.Set as Set                 ( Set(..), fromList, member)
+import safe qualified Data.Set as Set       ( Set(..), member, fromList )
+import safe Data.Traversable                ( Traversable(..) )
 import safe Data.Time.Calendar              ( Day
                                             , DayOfWeek
                                             , DayOfMonth
@@ -118,7 +136,7 @@ import safe Data.Time.Calendar.Quarter      ( QuarterOfYear
                                             , dayQuarter )
 import safe Data.Text                       ( pack, Text )
 import safe Data.Tuple                      ( fst, snd, uncurry, curry )
-
+import safe Data.Tuple.Curry                ( curryN, uncurryN )
 import safe IntervalAlgebra
 import safe IntervalAlgebra.IntervalUtilities
 import safe IntervalAlgebra.PairedInterval
@@ -126,3 +144,15 @@ import safe IntervalAlgebra.PairedInterval
 import safe Witherable                      ( Filterable(filter), Witherable(..) )
 import safe Flow                            ( (!>), (.>), (<!), (<.), (<|), (|>) )
 import Safe                                 ( headMay, lastMay )
+-- import Data.Vector.Fusion.Bundle (scanl1)
+-- import Data.Map (fromList)
+-- import qualified Data.Map as Set
+
+setFromList :: (Ord a) =>  [a] -> Set.Set a
+setFromList = Set.fromList
+
+mapFromList :: (Ord k) => [(k, a)] -> M.Map k a
+mapFromList = M.fromList
+
+mapToList :: (Ord k) =>  M.Map k a -> [(k, a)] 
+mapToList = M.toList