+++ src/Hasklepias/Templates/Features/Enrollment.hs
@@ -11,195 +11,234 @@ Maintainer  : bsaul@novisci.com
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
 
-module Hasklepias.Templates.Features.Enrollment (
-    defIsEnrolled 
-  , defContinuousEnrollment
-  , defEnrollmentTests
-) where
+module Hasklepias.Templates.Features.Enrollment
+  ( buildIsEnrolled
+  , buildContinuousEnrollment
+  , buildEnrollmentTests
+  ) where
 
-import Control.Applicative              ( Applicative(..) )
-import GHC.Int                          ( Int )
-import GHC.TypeLits                     ( KnownSymbol )
-import GHC.Show                         ( Show )
-import Flow                             ( (|>), (.>) )
-import IntervalAlgebra
-import IntervalAlgebra.IntervalUtilities ( combineIntervals )
-import IntervalAlgebra.PairedInterval   ( intervals )
-import Witherable                       ( Witherable, filter )
-import Data.Eq                          ( Eq )
-import Data.Functor.Contravariant       ( Predicate(..) )
-import Data.Foldable                    (Foldable(..), any)
-import Data.Function                    ( ($), (.) )
-import Data.Functor                     ( Functor(.. ) )
-import Data.Maybe                       ( Maybe )
-import Data.Monoid                      ( Monoid(..) )
-import Data.Text                        ( Text )
-import Data.Tuple                       ( uncurry )                     
-import Test.Tasty                       ( testGroup, TestName, TestTree )
-import Test.Tasty.HUnit                 ( testCase )
-
-import EventData                        ( Event
-                                        , Domain(..)
-                                        , EnrollmentFacts(..)
-                                        , event
-                                        , context
-                                        , isEnrollmentEvent
-                                        )
-import Features.Compose                 ( Feature
-                                        , Definition(..)
-                                        , Define(..)
-                                        , Eval(..)
-                                        , makeFeature )
-import Hasklepias.FeatureEvents         ( allGapsWithinLessThanDuration
-                                        )
-import Hasklepias.Templates.TestUtilities
-                                        ( makeAssertion
-                                        , TemplateTestCase(..) )
-import Hasklepias.Misc                  ( F )
-import Cohort.Index                     ( Index, makeIndex )
-
-import Cohort.AssessmentIntervals       ( makeBaselineFromIndex 
-                                        , AssessmentInterval
-                                        )
-import Cohort.Criteria                  ( Status(..), includeIf )
-
+import           Cohort
+import           EventData                  
+import           Features
+import           Hasklepias.Misc                ( F )
+import           Hasklepias.FeatureEvents
+import           Hasklepias.Templates.TestUtilities
+import           Hasklepias.Reexports
+import           Hasklepias.ReexportsUnsafe
 
 {-| Is Enrolled
 
 TODO: describe this
 
 -}
-defIsEnrolled ::
-  ( Intervallic i0 a
-  , Monoid (container (Interval a))
-  , Applicative container
-  , Witherable container) =>
+buildIsEnrolled
+  :: ( Intervallic i0 a
+     , Monoid (container (Interval a))
+     , Applicative container
+     , Witherable container
+     )
+  => 
   Predicate (Event a) -- ^ The predicate to filter to Enrollment events (e.g. 'FeatureEvents.isEnrollment')
   -> Definition
-  (   Feature indexName  (Index i0 a)
-   -> Feature eventsName (container (Event a))
-   -> Feature varName     Status )
-defIsEnrolled predicate =
-  define
-      (\index ->
-           filter (getPredicate predicate)
-        .> combineIntervals
-        .> any (concur index)
-        .> includeIf
-      )
+       (  Feature indexName (Index i0 a)
+       -> Feature eventsName (container (Event a))
+       -> Feature varName Status
+       )
+buildIsEnrolled predicate = define
+  (\index ->
+    filter (getPredicate predicate)
+      .> combineIntervals
+      .> any (concur index)
+      .> includeIf
+  )
 
-makeIsEnrolledTestInputs :: (IntervalSizeable a b) =>
-     TestName
-  -> b 
-  -> a
+makeIsEnrolledTestInputs
+  :: (Integral b, IntervalSizeable a b)
+  => TestName
+  -> Predicate (Event a)
+  -> (a, a)
   -> [Event a]
   -> Status
-  -> TemplateTestCase (F "index" (Index Interval a), F "events" [Event a]) Status
-makeIsEnrolledTestInputs name dur bgn e s = 
-  MkTemplateTestCase name (pure (makeIndex $ beginerval dur bgn), pure e) (pure s)
+  -> TestCase
+       (F "index" (Index Interval a), F "events" [Event a])
+       Status
+       (Predicate (Event a))
+makeIsEnrolledTestInputs name buildArgs intrvl e s = MkTestCase
+  buildArgs
+  name
+  (pure (makeIndex $ readIntervalSafe intrvl), pure e)
+  (pure s)
 
-makeEnrollmentEvent :: (IntervalSizeable a b) => b -> a -> Event a
-makeEnrollmentEvent dur bgn = 
-  event (beginerval dur bgn) (context ( Enrollment (EnrollmentFacts ())) mempty)
 
-defIsEnrolledTestCases :: [TemplateTestCase
-   (F "index" (Index Interval Int), F "events" [Event Int]) Status]
-defIsEnrolledTestCases = [ 
-      f "Exclude if no events" 1 (0::Int) []   Exclude
-    , f "Exclude if only interval meets" 1 (0::Int) [g 5 1]  Exclude
-    , f "Include if concurring interval" 1 (0::Int) [g 5 (-1)] Include
-    , f "Include if concurring interval" 1 (0::Int) [g 2 (-1), g 5 1]  Include 
-  ] where f = makeIsEnrolledTestInputs
-          g = makeEnrollmentEvent
+buildIsEnrolledTestCases
+  :: [ TestCase
+         (F "index" (Index Interval Int), F "events" [Event Int])
+         Status
+         (Predicate (Event Int))
+     ]
+buildIsEnrolledTestCases =
+  [ f "Exclude if no events" isEnrollmentEvent (0, 1) [] Exclude
+  , f "Exclude if only interval meets"
+      isEnrollmentEvent
+      (0, 1)
+      [g (1, 6)]
+      Exclude
+  , f "Include if concurring interval"
+      isEnrollmentEvent
+      (0, 1)
+      [g (-1, 4)]
+      Include
+  , f "Include if concurring interval"
+      isEnrollmentEvent
+      (0, 1)
+      [g (-1, 1), g (1, 4)]
+      Include
+  ] where
+  f = makeIsEnrolledTestInputs
+  g = makeEnrollmentEvent
 
-defIsEnrolledTests :: TestTree
-defIsEnrolledTests = testGroup "Tests of isEnrolled template"
-     ( fmap (\x -> testCase (getTestName x) (makeAssertion x (defIsEnrolled isEnrollmentEvent)) )
-       defIsEnrolledTestCases )
+buildIsEnrolledTests :: TestTree
+buildIsEnrolledTests = testGroup
+  "Tests of isEnrolled template"
+  (fmap
+    (\x -> testCase (getTestName x)
+                    (makeAssertion x (buildIsEnrolled (getBuilderArgs x)))
+    )
+    buildIsEnrolledTestCases
+  )
 
 
 {-| Continuous Enrollment 
 
 TODO: describe this
+
 -}
-defContinuousEnrollment ::
-  ( Monoid (container (Interval a))
-  , Monoid (container (Maybe (Interval a)))
-  , Applicative container
-  , Witherable container
-  , IntervalSizeable a b) =>
-    (Index i0 a -> AssessmentInterval a) -- ^ function which maps index interval to interval in which to assess enrollment
+buildContinuousEnrollment
+  :: ( Monoid (container (Interval a))
+     , Monoid (container (Maybe (Interval a)))
+     , Applicative container
+     , Witherable container
+     , IntervalSizeable a b
+     )
+  => (Index i0 a -> AssessmentInterval a) -- ^ function which maps index interval to interval in which to assess enrollment
   -> Predicate (Event a)  -- ^ The predicate to filter to Enrollment events (e.g. 'FeatureEvents.isEnrollment')
   -> b  -- ^ duration of allowable gap between enrollment intervals
   -> Definition
-  (   Feature indexName  (Index i0 a)
-   -> Feature eventsName (container (Event a))
-   -> Feature prevName    Status
-   -> Feature varName     Status )
-defContinuousEnrollment makeAssessmentInterval predicate allowableGap =
-  define
-    (\index events prevStatus ->
-      case prevStatus of
-        Exclude -> Exclude
-        Include -> includeIf
-          ( allGapsWithinLessThanDuration
-                allowableGap
-                (makeAssessmentInterval index)
-                (combineIntervals $ filter (getPredicate predicate) events))
-    )
+       (  Feature indexName (Index i0 a)
+       -> Feature eventsName (container (Event a))
+       -> Feature prevName Status
+       -> Feature varName Status
+       )
+buildContinuousEnrollment makeAssessmentInterval predicate allowableGap = define
+  (\index events prevStatus -> case prevStatus of
+    Exclude -> Exclude
+    Include -> includeIf
+      (allGapsWithinLessThanDuration
+        allowableGap
+        (makeAssessmentInterval index)
+        (combineIntervals $ filter (getPredicate predicate) events)
+      )
+  )
 
-makeContinuousEnrollmentTestInputs :: (IntervalSizeable a b) =>
-     TestName
-  -> b 
-  -> a
+
+type ContEnrollArgs
+  = (Index Interval Int -> AssessmentInterval Int, Predicate (Event Int), Int)
+
+makeContinuousEnrollmentTestInputs
+  :: (Integral b, IntervalSizeable a b)
+  => TestName
+  -> ContEnrollArgs
+  -> (a, a)
   -> [Event a]
   -> Status
   -> Status
-  -> TemplateTestCase (F "index" (Index Interval a), F "events" [Event a], F "prev" Status) Status
-makeContinuousEnrollmentTestInputs name dur bgn e prev s = 
-  MkTemplateTestCase name (pure (makeIndex $ beginerval dur bgn), pure e, pure prev) (pure s)
+  -> TestCase
+       ( F "index" (Index Interval a)
+       , F "events" [Event a]
+       , F "prev" Status
+       )
+       Status
+       ContEnrollArgs
+makeContinuousEnrollmentTestInputs name buildArgs intrvl e prev s = MkTestCase
+  buildArgs
+  name
+  (pure (makeIndex (readIntervalSafe intrvl)), pure e, pure prev)
+  (pure s)
 
-defContinuousEnrollmentTestCases :: [TemplateTestCase
-   (F "index" (Index Interval Int), F "events" [Event Int], F "prev" Status) Status]
-defContinuousEnrollmentTestCases = [ 
-      f "Exclude if previously excluded" 1 (0::Int) []  Exclude  Exclude
-    , f "Exclude if no events" 1 (0::Int) []  Include  Exclude
+commonArgs
+  :: (Index Interval Int -> AssessmentInterval Int, Predicate (Event a), Int)
+commonArgs = (makeBaselineFromIndex 10, isEnrollmentEvent, 3)
+
+buildContinuousEnrollmentTestCases
+  :: [ TestCase
+         ( F "index" (Index Interval Int)
+         , F "events" [Event Int]
+         , F "prev" Status
+         )
+         Status
+         ContEnrollArgs
+     ]
+buildContinuousEnrollmentTestCases =
+  [ f "Exclude if previously excluded" commonArgs (0, 1) [] Exclude Exclude
+  , f "Exclude if no events"           commonArgs (0, 1) [] Include Exclude
+  , f "Exclude if gap >= 3"
+      commonArgs
+      (10, 11)
+      [g (1, 4), g (9, 12)]
+      Include
+      Exclude
       {-
                   -           <- Index
          ----------           <- Baseline
          ---     ---          <- Enrollment
         |--------------|
       -}
-    , f "Exclude if gap >= 3" 1 (10::Int) [g 3 1, g 3 9]  Include Exclude
+  , f "Exclude if gap >= 3" commonArgs (10, 11) [g (1, 7)]  Include Exclude
       {-
                   -           <- Index
         ----------            <- Baseline
          ------               <- Enrollment
         |--------------|
       -}
-    , f "Exclude if gap >= 3" 1 (10::Int) [g 6 1]   Include Exclude
+  , f "Exclude if gap >= 3" commonArgs (10, 11) [g (6, 13)] Include Exclude
         {-
                   -           <- Index
          ----------           <- Baseline
               -------         <- Enrollment
         |--------------|
       -}
-    , f "Exclude if gap >= 3" 1 (10::Int) [g 7 6]   Include Exclude
+  , f "Include if gaps less than 3"
+      commonArgs
+      (10, 11)
+      [g (1, 3), g (5, 12)]
+      Include
+      Include
       {-
                   -           <- Index
          ----------           <- Baseline
          --  -------          <- Enrollment
         |--------------|
       -}
-    , f "Include if gaps less than 3" 1 (10::Int) [g 2 1, g 7 5]  Include Include
+  , f "Include if gaps less than 3"
+      commonArgs
+      (10, 11)
+      [g (2, 9)]
+      Include
+      Include
       {-
                   -           <- Index
          ----------           <- Baseline
           -------             <- Enrollment
         |--------------|
       -}
-    , f "Include if gaps less than 3" 1 (10::Int) [g 7 2]  Include Include
+  , f "Include if gaps less than 3"
+      commonArgs
+      (10, 11)
+      [g (1, 6), g (4, 8)]
+      Include
+      Include
         {-
                   -           <- Index
          ----------           <- Baseline
@@ -207,15 +246,24 @@ defContinuousEnrollmentTestCases = [
              ----
         |--------------|
       -}
-    , f "Include if gaps less than 3" 1 (10::Int) [g 5 1, g 4 4]  Include Include
-  ] where f = makeContinuousEnrollmentTestInputs
-          g = makeEnrollmentEvent
+  ] where
+  f = makeContinuousEnrollmentTestInputs
+  g = makeEnrollmentEvent
 
-defContinuousEnrollmentTests :: TestTree
-defContinuousEnrollmentTests = testGroup "Tests of continuous enrollment template"
-     ( fmap (\x -> testCase (getTestName x) 
-          (makeAssertion x (defContinuousEnrollment (makeBaselineFromIndex 10) isEnrollmentEvent 3)) )
-           defContinuousEnrollmentTestCases )
+buildContinuousEnrollmentTests :: TestTree
+buildContinuousEnrollmentTests = testGroup
+  "Tests of continuous enrollment template"
+  (fmap
+    (\x -> testCase
+      (getTestName x)
+      (makeAssertion
+        x
+        (buildContinuousEnrollment (makeBaselineFromIndex 10) isEnrollmentEvent 3)
+      )
+    )
+    buildContinuousEnrollmentTestCases
+  )
 
-defEnrollmentTests :: TestTree
-defEnrollmentTests = testGroup "" [defIsEnrolledTests, defContinuousEnrollmentTests]
+buildEnrollmentTests :: TestTree
+buildEnrollmentTests =
+  testGroup "" [buildIsEnrolledTests, buildContinuousEnrollmentTests]